ml_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
ta-lib==0.4.28
torch==2.1.1
transformers==4.35.2
scikit-learn==1.3.2
pymongo==4.6.0
python-binance==1.0.19
yfinance==0.2.31
redis==5.0.1
pydantic==2.5.2
pytest==7.4.3
python-jose==3.3.0
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6 